Extract LanguageInformation section to own component
