+++ frontend/talentsearch/src/js/components/profile/ProfilePage/ProfilePage.tsx
@@ -15,13 +15,9 @@ import { Link } from "@common/components";
 import TableOfContents from "@common/components/TableOfContents";
 import commonMessages from "@common/messages/commonMessages";
 import { imageUrl } from "@common/helpers/router";
-import { getLocale } from "@common/helpers/localize";
 import {
-  getLanguage,
   getOperationalRequirement,
   getWorkRegion,
-  getProvinceOrTerritory,
-  getLanguageProficiency,
   womanLocalized,
   indigenousLocalized,
   minorityLocalized,
@@ -34,7 +30,6 @@ import { unpackMaybes } from "@common/helpers/formUtils";
 import TALENTSEARCH_APP_DIR from "../../../talentSearchConstants";
 import { useApplicantProfileRoutes } from "../../../applicantProfileRoutes";
 import {
-  BilingualEvaluation,
   useGetMeQuery,
   User,
   GetMeQuery,
@@ -44,6 +39,7 @@ import {
 import MyStatusApi from "../../myStatusForm/MyStatusForm";
 import CandidatePoolsSection from "./CandidatePoolsSection";
 import AboutMeSection from "./AboutMeSection";
+import LanguageInformationSection from "./LanguageInformationSection";
 
 export interface ProfilePageProps {
   profileDataInput: User;
@@ -60,19 +56,6 @@ export const ProfileForm: React.FC<ProfilePageProps> = ({
   const {
     firstName,
     lastName,
-    email,
-    telephone,
-    preferredLang,
-    currentProvince,
-    currentCity,
-    lookingForEnglish,
-    lookingForFrench,
-    lookingForBilingual,
-    bilingualEvaluation,
-    comprehensionLevel,
-    writtenLevel,
-    verbalLevel,
-    estimatedLanguageAbility,
     isGovEmployee,
     govEmployeeType,
     interestedInLaterOrSecondment,
@@ -276,192 +259,10 @@ export const ProfileForm: React.FC<ProfilePageProps> = ({
                 })}
               </Link>
             </div>
-            <div
-              data-h2-bg-color="b(lightgray)"
-              data-h2-padding="b(all, m)"
-              data-h2-radius="b(s)"
-            >
-              {lookingForEnglish && !lookingForFrench && !lookingForBilingual && (
-                <p>
-                  {intl.formatMessage({
-                    defaultMessage: "Interested in:",
-                    description: "Interested in label and colon",
-                  })}{" "}
-                  <span data-h2-font-weight="b(700)">
-                    {intl.formatMessage({
-                      defaultMessage: "English positions",
-                      description: "English Positions message",
-                    })}
-                  </span>
-                </p>
-              )}
-              {!lookingForEnglish && lookingForFrench && !lookingForBilingual && (
-                <p>
-                  {intl.formatMessage({
-                    defaultMessage: "Interested in:",
-                    description: "Interested in label and colon",
-                  })}{" "}
-                  <span data-h2-font-weight="b(700)">
-                    {intl.formatMessage({
-                      defaultMessage: "French positions",
-                      description: "French Positions message",
-                    })}
-                  </span>
-                </p>
-              )}
-              {lookingForEnglish && lookingForFrench && !lookingForBilingual && (
-                <p>
-                  {intl.formatMessage({
-                    defaultMessage: "Interested in:",
-                    description: "Interested in label and colon",
-                  })}{" "}
-                  <span data-h2-font-weight="b(700)">
-                    {intl.formatMessage({
-                      defaultMessage: "English or French positions",
-                      description: "English or French Positions message",
-                    })}
-                  </span>
-                </p>
-              )}
-              {lookingForBilingual && (
-                <p>
-                  {intl.formatMessage({
-                    defaultMessage: "Interested in:",
-                    description: "Interested in label and colon",
-                  })}{" "}
-                  <span data-h2-font-weight="b(700)">
-                    {intl.formatMessage({
-                      defaultMessage:
-                        "Bilingual positions (English and French)",
-                      description: "Bilingual Positions message",
-                    })}
-                  </span>
-                </p>
-              )}
-              {bilingualEvaluation === BilingualEvaluation.CompletedEnglish && (
-                <p>
-                  {intl.formatMessage({
-                    defaultMessage: "Completed an official GoC evaluation:",
-                    description:
-                      "Completed a government of canada abbreviation evaluation label and colon",
-                  })}{" "}
-                  <span data-h2-font-weight="b(700)">
-                    {intl.formatMessage({
-                      defaultMessage: "Yes, completed ENGLISH evaluation",
-                      description: "Completed an English language evaluation",
-                    })}
-                  </span>
-                </p>
-              )}
-              {bilingualEvaluation === BilingualEvaluation.CompletedFrench && (
-                <p>
-                  {intl.formatMessage({
-                    defaultMessage: "Completed an official GoC evaluation:",
-                    description:
-                      "Completed a government of canada abbreviation evaluation label and colon",
-                  })}{" "}
-                  <span data-h2-font-weight="b(700)">
-                    {intl.formatMessage({
-                      defaultMessage: "Yes, completed FRENCH evaluation",
-                      description: "Completed a French language evaluation",
-                    })}
-                  </span>
-                </p>
-              )}
-              {bilingualEvaluation === BilingualEvaluation.NotCompleted && (
-                <p>
-                  {intl.formatMessage({
-                    defaultMessage: "Completed an official GoC evaluation:",
-                    description:
-                      "Completed a government of canada abbreviation evaluation label and colon",
-                  })}{" "}
-                  <span data-h2-font-weight="b(700)">
-                    {intl.formatMessage({
-                      defaultMessage: "No",
-                      description:
-                        "No, did not completed a language evaluation",
-                    })}
-                  </span>
-                </p>
-              )}
-              {(bilingualEvaluation === BilingualEvaluation.CompletedEnglish ||
-                bilingualEvaluation ===
-                  BilingualEvaluation.CompletedFrench) && (
-                <p>
-                  {intl.formatMessage({
-                    defaultMessage:
-                      "Second language level (Comprehension, Written, Verbal):",
-                    description:
-                      "Evaluation results for second language, results in that order followed by a colon",
-                  })}{" "}
-                  <span data-h2-font-weight="b(700)">
-                    {comprehensionLevel}, {writtenLevel}, {verbalLevel}
-                  </span>
-                </p>
-              )}
-              {bilingualEvaluation === BilingualEvaluation.NotCompleted &&
-                !!estimatedLanguageAbility && (
-                  <p>
-                    {intl.formatMessage({
-                      defaultMessage: "Second language level:",
-                      description:
-                        "Estimated skill in second language, followed by a colon",
-                    })}{" "}
-                    <span data-h2-font-weight="b(700)">
-                      {estimatedLanguageAbility
-                        ? getLanguageProficiency(estimatedLanguageAbility)
-                            .defaultMessage
-                        : ""}
-                    </span>
-                  </p>
-                )}
-              {!lookingForEnglish &&
-                !lookingForFrench &&
-                !lookingForBilingual &&
-                !bilingualEvaluation && (
-                  <p>
-                    {intl.formatMessage({
-                      defaultMessage:
-                        "You haven't added any information here yet.",
-                      description:
-                        "Message for when no data exists for the section",
-                    })}
-                  </p>
-                )}
-              {((!lookingForEnglish &&
-                !lookingForFrench &&
-                !lookingForBilingual) ||
-                (lookingForBilingual &&
-                  (!bilingualEvaluation ||
-                    ((bilingualEvaluation ===
-                      BilingualEvaluation.CompletedEnglish ||
-                      bilingualEvaluation ===
-                        BilingualEvaluation.CompletedFrench) &&
-                      (!comprehensionLevel ||
-                        !writtenLevel ||
-                        !verbalLevel))))) && (
-                <p>
-                  {intl.formatMessage(
-                    {
-                      defaultMessage:
-                        "There are <redText>required</redText> fields missing.",
-                      description:
-                        "Message that there are required fields missing. Please ignore things in <> tags.",
-                    },
-                    {
-                      redText,
-                    },
-                  )}{" "}
-                  <a href={paths.languageInformation()}>
-                    {intl.formatMessage({
-                      defaultMessage: "Click here to get started.",
-                      description:
-                        "Message to click on the words to begin something",
-                    })}
-                  </a>
-                </p>
-              )}
-            </div>
+            <LanguageInformationSection
+              applicant={profileDataInput}
+              editPath={paths.languageInformation()}
+            />
           </TableOfContents.Section>
           <TableOfContents.Section id="gov-info-section">
             <div style={{ display: "flex", alignItems: "baseline" }}>
